requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
ansi2html==1.8.0
async-generator==1.10
attrs==23.1.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
coverage==7.2.5
decorator==5.1.1
docutils==0.19
exceptiongroup==1.1.1
execnet==1.9.0
future-fstrings==1.2.0
h11==0.14.0
htmlmin==0.1.12
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.20.1
networkx==3.1
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PySocks==1.7.1
pytest==7.3.1
pytest-cov==4.0.0
pytest-depends==1.0.1
pytest-dotenv==0.5.2
pytest-html==3.2.0
pytest-jira==0.3.18
pytest-metadata==2.0.4
pytest-reporter==0.5.2
pytest-xdist==3.2.1
python-dotenv==1.0.0
requests==2.29.0
retry==0.9.2
selenium==4.9.0
setuptools==69.1.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tblib==1.7.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.10.0
urllib3==1.26.15
webdriver-manager==4.0.0
wsproto==1.2.0
